Integration tests for data loading + CI attempt[skip ci]
