+++ .github/workflows/build_macos.yml
@@ -72,8 +72,8 @@ jobs:
     - name: Make installer
       run: |
         git clone https://github.com/dbouget/quickpkg.git
-        quickpkg/quickpkg dist/Raidionics.app --output Raidionics-1.2.4-macOS.pkg
-        cp -r Raidionics-1.2.4-macOS.pkg dist/Raidionics-1.2.4-macOS-x86_64.pkg
+        quickpkg/quickpkg dist/Raidionics.app --output Raidionics-1.3.0-macOS.pkg
+        cp -r Raidionics-1.3.0-macOS.pkg dist/Raidionics-1.3.0-macOS-x86_64.pkg
 
     - name: Upload package
       uses: actions/upload-artifact@v4
