+++ .github/workflows/build_macos_arm.yml
@@ -67,8 +67,8 @@ jobs:
     - name: Make installer
       run: |
         git clone https://github.com/dbouget/quickpkg.git
-        quickpkg/quickpkg dist/Raidionics.app --output Raidionics-1.2.4-macOS.pkg
-        cp -r Raidionics-1.2.4-macOS.pkg dist/Raidionics-1.2.4-macOS-arm64.pkg
+        quickpkg/quickpkg dist/Raidionics.app --output Raidionics-1.3.0-macOS.pkg
+        cp -r Raidionics-1.3.0-macOS.pkg dist/Raidionics-1.3.0-macOS-arm64.pkg
 
     - name: Upload package
       uses: actions/upload-artifact@v4
