+++ .github/workflows/build_ubuntu.yml
@@ -106,7 +106,7 @@ jobs:
       env:
         DISPLAY: ':1'
       run: |
-          pip install pytest
+          pip install pytest-qt
           pytest ${{github.workspace}}/integration_tests
 
     - name: Make installer
@@ -115,7 +115,7 @@ jobs:
         cp -r dist/Raidionics assets/Raidionics_ubuntu/usr/local/bin
         dpkg-deb --build --root-owner-group assets/Raidionics_ubuntu
         ls -la
-        cp -r assets/Raidionics_ubuntu.deb dist/Raidionics-1.2.4-ubuntu.deb
+        cp -r assets/Raidionics_ubuntu.deb dist/Raidionics-1.3.0-ubuntu.deb
 
     - name: Upload package
       uses: actions/upload-artifact@v4
