+++ .github/workflows/build_windows.yml
@@ -62,7 +62,7 @@ jobs:
     - name: Make installer
       run: |
         makensis.exe assets/Raidionics.nsi
-        cp -r assets/Raidionics-1.2.4-win.exe dist/Raidionics-1.2.4-win.exe
+        cp -r assets/Raidionics-1.3.0-win.exe dist/Raidionics-1.3.0-win.exe
 
     - name: Upload package
       uses: actions/upload-artifact@v4
