+++ assets/Raidionics.nsi
@@ -1,8 +1,8 @@
 !define APP_NAME "Raidionics"
 !define COMP_NAME "SINTEF"
-!define VERSION "1.2.2"
+!define VERSION "1.3.0"
 !define DESCRIPTION "Application"
-!define INSTALLER_NAME "Raidionics-1.2.4-win.exe"
+!define INSTALLER_NAME "Raidionics-1.3.0-win.exe"
 !define MAIN_APP_EXE "Raidionics.exe"
 !define INSTALL_TYPE "SetShellVarContext current"
 !define REG_ROOT "HKLM"
