+++ assets/Raidionics_ubuntu/DEBIAN/control
@@ -1,5 +1,5 @@
 Package: Raidionics
-Version: 1.2.4
+Version: 1.3.0
 Architecture: i386
 Maintainer: David Bouget <david.bouget@sintef.no>
 Description: Raidionics—Reporting and Data System.
